+++ tools/cooja/java/se/sics/cooja/TimeEvent.java
@@ -43,18 +43,18 @@ public abstract class TimeEvent {
   EventQueue scheduledIn = null;
   String name;
 
-  protected int time;
+  protected long time;
 
-  public TimeEvent(int time) {
+  public TimeEvent(long time) {
     this.time = time;
   }
 
-  public TimeEvent(int time, String name) {
+  public TimeEvent(long time, String name) {
     this.time = time;
     this.name = name;
   }
 
-  public final int getTime() {
+  public final long getTime() {
     return time;
   }
 
@@ -69,7 +69,7 @@ public abstract class TimeEvent {
     return false;
   }
 
-  public abstract void execute(int t);
+  public abstract void execute(long t);
 
   public String getShort() {
     return "" + time + (name != null ? ": " + name : "");
